+++ lib/main.directories.g.dart
@@ -13,14 +13,16 @@ import 'package:pactus_gui_widgetbook/src/features/widgets/buttons/adaptive_prim
     as _i2;
 import 'package:pactus_gui_widgetbook/src/features/widgets/buttons/adaptive_secondary_button/adaptive_secondary_button_usecase.dart'
     as _i3;
-import 'package:pactus_gui_widgetbook/src/features/widgets/chip_text_box/chip_text_box_use_case.dart'
+import 'package:pactus_gui_widgetbook/src/features/widgets/buttons/adaptive_text_button/adaptive_text_button_usecase.dart'
     as _i4;
-import 'package:pactus_gui_widgetbook/src/features/widgets/counter/counter_usecase.dart'
+import 'package:pactus_gui_widgetbook/src/features/widgets/chip_text_box/chip_text_box_use_case.dart'
     as _i5;
-import 'package:pactus_gui_widgetbook/src/features/widgets/custom_button/custom_button_usecase.dart'
+import 'package:pactus_gui_widgetbook/src/features/widgets/counter/counter_usecase.dart'
     as _i6;
-import 'package:pactus_gui_widgetbook/src/features/widgets/custom_text/custom_text_usecase.dart'
+import 'package:pactus_gui_widgetbook/src/features/widgets/custom_button/custom_button_usecase.dart'
     as _i7;
+import 'package:pactus_gui_widgetbook/src/features/widgets/custom_text/custom_text_usecase.dart'
+    as _i8;
 import 'package:widgetbook/widgetbook.dart' as _i1;
 
 final directories = <_i1.WidgetbookNode>[
@@ -57,6 +59,18 @@ final directories = <_i1.WidgetbookNode>[
                   )
                 ],
               ),
+              _i1.WidgetbookFolder(
+                name: 'adaptive_text_button',
+                children: [
+                  _i1.WidgetbookLeafComponent(
+                    name: 'AdaptiveTextButton',
+                    useCase: _i1.WidgetbookUseCase(
+                      name: 'Adaptive Text Button with Knobs',
+                      builder: _i4.adaptiveTextButtonUseCase,
+                    ),
+                  )
+                ],
+              ),
             ],
           ),
           _i1.WidgetbookFolder(
@@ -66,7 +80,7 @@ final directories = <_i1.WidgetbookNode>[
                 name: 'ChipTextBox',
                 useCase: _i1.WidgetbookUseCase(
                   name: 'ChipTextBox Customization',
-                  builder: _i4.chipTextBoxUseCase,
+                  builder: _i5.chipTextBoxUseCase,
                 ),
               )
             ],
@@ -78,7 +92,7 @@ final directories = <_i1.WidgetbookNode>[
                 name: 'CounterWidget',
                 useCase: _i1.WidgetbookUseCase(
                   name: 'Counter with Knobs',
-                  builder: _i5.counterWidgetUseCase,
+                  builder: _i6.counterWidgetUseCase,
                 ),
               )
             ],
@@ -90,7 +104,7 @@ final directories = <_i1.WidgetbookNode>[
                 name: 'CustomButtonWidget',
                 useCase: _i1.WidgetbookUseCase(
                   name: 'Custom Button with Knobs',
-                  builder: _i6.customButtonWidgetUseCase,
+                  builder: _i7.customButtonWidgetUseCase,
                 ),
               )
             ],
@@ -102,7 +116,7 @@ final directories = <_i1.WidgetbookNode>[
                 name: 'CustomTextWidget',
                 useCase: _i1.WidgetbookUseCase(
                   name: 'Select the text to display from the dropdown',
-                  builder: _i7.customTextWidgetUseCase,
+                  builder: _i8.customTextWidgetUseCase,
                 ),
               )
             ],