+++ lib/src/features/widgets/buttons/adaptive_primary_button/adaptive_primary_button.dart
@@ -16,45 +16,26 @@ import 'package:pactus_gui_widgetbook/src/features/widgets/buttons/core/utils/me
 ///   - Defines the state of the button (e.g., loading, success, failure).
 ///   - Determines if the button should be disabled.
 ///
-/// - **[onPressed]** (`Function()?`):
-///   - Callback function triggered when the button is pressed.
-///   - Disabled if `requestState` is `loading`.
+/// The `AdaptivePrimaryButton` is a customizable button widget designed to handle various button types based on the content (text, icons, or both). It adapts its layout and style based on different states such as loading, success, or failure. This button is particularly useful for situations where the button needs to display dynamic content, including text and icons, with adaptive behavior for different states and interactions.
 ///
-/// - **[suffixIcon]** (`IconData?`):
-///   - An optional suffix icon displayed at the end of the button.
+/// ### Constructor Parameters:
+/// - **[requestState]** (`RequestStateEnum`): Represents the current state of the request. It controls the button's loading state and whether the button can be pressed.
+/// - **[onPressed]** (`Function()?`): The callback function to be invoked when the button is pressed.
+/// - **[suffixIcon]** (`IconData?`): The icon to be displayed at the end of the button (after the title).
+/// - **[prefixIcon]** (`IconData?`): The icon to be displayed at the beginning of the button (before the title).
+/// - **[title]** (`String?`): The text displayed on the button.
+/// - **[buttonType]** (`ButtonTypeEnum`): Specifies the type of button layout, including options like title-only, icon-and-title, and others.
+/// - **[baseIcon]** (`IconData?`): The main icon to be displayed when only the icon is needed (used in the `iconOnly` button type).
+/// - **[paddingSize]** (`PaddingSizeEnum`): Defines the horizontal padding size for the button content. It determines the spacing inside the button around the text and icons.
+/// - **[isOutlined]** (`bool`): If `true`, the button will have an outline, otherwise, it will have a solid fill.
 ///
-/// - **[prefixIcon]** (`IconData?`):
-///   - An optional prefix icon displayed at the start of the button.
-///
-/// - **[title]** (`String?`):
-///   - The text label of the button.
-///   - Optional and used in applicable button configurations.
-///
-/// - **[buttonType]** (`ButtonTypeEnum`):
-///   - Specifies the type of button (e.g., title-only, icon-only, both).
-///
-/// - **[baseIcon]** (`IconData?`):
-///   - The primary icon used when the button is of type `iconOnly`.
-///
-/// - **[paddingSize]** (`PaddingSizeEnum`):
-///   - Determines the padding size of the button.
-///
-/// ### Factory Methods:
-///
-/// - **[AdaptivePrimaryButton.createTitleOnly]**:
-///   - Creates a button with only a title.
-///
-/// - **[AdaptivePrimaryButton.createIconAndTitle]**:
-///   - Creates a button with a title and a leading icon.
-///
-/// - **[AdaptivePrimaryButton.createTitleAndIcon]**:
-///   - Creates a button with a title and a trailing icon.
-///
-/// - **[AdaptivePrimaryButton.createIconTitleAndIcon]**:
-///   - Creates a button with both leading and trailing icons along with a title.
-///
-/// - **[AdaptivePrimaryButton.createIconOnly]**:
-///   - Creates a button that contains only an icon.
+/// ### Factory Constructors:
+/// These factory methods create `AdaptivePrimaryButton` instances with different button types:
+/// - **[createTitleOnly]**: A button with only text (title).
+/// - **[createIconAndTitle]**: A button with an icon before the text.
+/// - **[createTitleAndIcon]**: A button with the text followed by an icon.
+/// - **[createIconTitleAndIcon]**: A button with an icon before and after the text.
+/// - **[createIconOnly]**: A button that only shows an icon.
 ///
 
 class AdaptivePrimaryButton extends StatelessWidget {